# tests-ts/contracts/fastapi-openapi.json
{
  "components": {
    "schemas": {
...
      "Body_start_copilot_prep_api_copilot_prep_post": {
        "properties": {
          "company": {
            "default": "",
            "title": "Company",
            "type": "string"
          },
          "jd_text": {
            "title": "Jd Text",
            "type": "string"
          },
          "position": {
            "default": "",
            "title": "Position",
            "type": "string"
          }
        },
        "required": [
          "jd_text"
        ],
        "title": "Body_start_copilot_prep_api_copilot_prep_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "TechSpar",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/copilot/prep": {
      "post": {
        "description": "启动 Copilot Prep Phase（后台异步执行）。",
        "operationId": "start_copilot_prep_api_copilot_prep_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_start_copilot_prep_api_copilot_prep_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Copilot Prep"
      }
    },
    "/api/copilot/prep/{prep_id}": {
      "delete": {
        "description": "删除一个 Copilot Prep 会话。",
        "operationId": "delete_copilot_prep_api_copilot_prep__prep_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "prep_id",
            "required": true,
            "schema": {
              "title": "Prep Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Copilot Prep"
      },
      "get": {
        "description": "查询 Copilot Prep 进度和结果。",
        "operationId": "get_copilot_prep_status_api_copilot_prep__prep_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "prep_id",
            "required": true,
            "schema": {
              "title": "Prep Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Copilot Prep Status"
      }
    },
    "/api/copilot/prep/{prep_id}/tree": {
      "get": {
        "description": "获取策略树（前端可视化用）。",
        "operationId": "get_copilot_strategy_tree_api_copilot_prep__prep_id__tree_get",
        "parameters": [
          {
            "in": "path",
            "name": "prep_id",
            "required": true,
            "schema": {
              "title": "Prep Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Copilot Strategy Tree"
      }
    },
    "/api/copilot/preps": {
      "get": {
        "description": "列出当前用户的所有 Copilot Prep 会话。",
        "operationId": "list_copilot_preps_api_copilot_preps_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Copilot Preps"
      }
    },
...
  }
}